+++ lib/extract_carbon_prices.rb
@@ -2,7 +2,7 @@ require_relative 'gdx'
 
 class ExtractCarbonPrices
 
-  SYMBOLS_TO_CHECK_FOR_PRICE = %w{ EQE_COMBAL EQG_COMBAL EQL_COMBAL EQE_COMNET EQL_COMNET EQG_COMNET }
+  SYMBOLS_TO_CHECK_FOR_PRICE = %w{ PAR_COMBALEM PAR_COMBALGM PAR_COMNETM PAR_COMPRDM }
 
   attr_accessor :gdx
   attr_accessor :scenario_name
@@ -23,7 +23,7 @@ class ExtractCarbonPrices
   attr_accessor :symbols
 
   def extract_all_symbols
-    @symbols = SYMBOLS_TO_CHECK_FOR_PRICE.map { |symbol| gdx.marginals(symbol) }
+    @symbols = SYMBOLS_TO_CHECK_FOR_PRICE.map { |symbol| gdx.symbol(symbol) }
   end
 
   def filter_carbon_prices
@@ -37,14 +37,26 @@ class ExtractCarbonPrices
   end
 
   def transform_carbon_prices
-    prices = Hash.new { |hash, new_key| hash[new_key] = Hash.new { |hash, new_key| hash[new_key] = 0.0 } }
+    prices = Hash.new { |hash, new_key| hash[new_key] = {} }
 
     symbols.each do |marginals|
       marginals.each do |marginal|
         commodity = marginal[:c]
         year = marginal[:allyear]
         value = marginal[:val] * -1000.0 # £M/ktCO2e to £/tCO2e / Negative because if loosening the constraint lowers the Objective value, then the carbon price must be positive
-        prices[commodity][year] = prices[commodity][year] + value
+	if prices[commodity][year]
+		if value == 0
+			# leave it alone, it will be a bigger absolute number
+		elsif value > 0
+           		prices[commodity][year] = [prices[commodity][year], value].max
+		elsif prices[commodity][year] < 0
+           		prices[commodity][year] = [prices[commodity][year], value].min
+		else
+			# leave it alone it will be a postitive number
+		end
+	else
+        	prices[commodity][year] = value
+	end
       end
     end
     { :name => scenario_name, :prices => prices }